rename cursor component and move it to a new file under canvas controls
