+++ editor/src/components/canvas/controls/new-canvas-controls.tsx
@@ -72,7 +72,7 @@ import { InsertionControls } from './insertion-plus-button'
 import { DistanceGuidelineControl } from './select-mode/distance-guideline-control'
 import { SceneLabelControl } from './select-mode/scene-label'
 import { PinLines } from './position-outline'
-import { EditorCursorComponent } from '../../editor/editor-component'
+import { CursorOverlay } from './select-mode/cursor-overlay'
 
 export const CanvasControlsContainerID = 'new-canvas-controls-container'
 
@@ -488,7 +488,7 @@ const NewCanvasControlsInner = (props: NewCanvasControlsInnerProps) => {
           )}
         </>,
       )}
-      <EditorCursorComponent />
+      <CursorOverlay />
     </div>
   )
 }
