+++ editor/src/components/editor/editor-component.tsx
@@ -456,37 +456,6 @@ export function EditorComponent(props: EditorProps) {
   )
 }
 
-export const EditorCursorComponent = React.memo(() => {
-  const cursor = useEditorState((store) => {
-    return Utils.defaultIfNull(store.editor.canvas.cursor, getCursorFromDragState(store.editor))
-  }, 'EditorCursorComponent cursor')
-
-  const styleProps = React.useMemo(() => {
-    let workingStyleProps: React.CSSProperties = {
-      position: 'fixed',
-      left: 0,
-      top: 0,
-      width: '100vw',
-      height: '100vh',
-      pointerEvents: 'none',
-      zIndex: 9999999,
-    }
-    if (cursor != null) {
-      workingStyleProps.cursor = cursor
-      workingStyleProps.pointerEvents = 'all'
-    }
-    return workingStyleProps
-  }, [cursor])
-  const portalDiv = document.getElementById('cursor-overlay-portal')
-  if (portalDiv == null) {
-    return null
-  }
-  return ReactDOM.createPortal(
-    <div key='cursor-area' id='cursor-overlay' style={styleProps} />,
-    portalDiv,
-  )
-})
-
 const ToastRenderer = React.memo(() => {
   const toasts = useEditorState((store) => store.editor.toasts, 'ToastRenderer')
